test(restaurant_registration): add reuse to scenario step functions
