+++ backend/tests/features/tests.feature
@@ -1,10 +1,8 @@
 Feature: Tests
 
 # API
-Scenario: Leitura de restaurantes do sistema
-    Given existe um restaurante cadastrado no sistema com os dados "Quentinha refeições" "123321222", email "email_adm_restaurante" e senha "senha_adm_restaurante"
-    And existe um restaurante cadastrado no sistema com os dados "Guloso Trincado" "40028922", email "email_adm_restaurante_2" e senha "senha_adm_restaurante_2"
-    When uma requisição GET é enviada para "/restaurants"
-    Then é retornada uma mensagem com o status "200"
-    And a mensagem contém "Quentinha refeições", "123321222", "email_adm_restaurante"
-    And a mensagem contém "Guloso Trincado", "40028922", "email_adm_restaurante_2"
+Scenario: Create a test
+    Given o TestRepository não tem um test com nome "test"
+    When uma requisição POST for enviada para "/api/tests" com o corpo da requisição sendo um JSON com o nome "test"
+    Then o status da resposta deve ser "200"
+    And o JSON da resposta deve conter o nome "test"
